+++ lib/mail/notify/delivery_method.rb
@@ -6,6 +6,8 @@ module Mail
       attr_accessor :settings, :response
 
       def initialize(settings)
+        raise ArgumentError, 'You must specify an API key' if settings[:api_key].blank?
+
         @settings = settings
       end